removed initial render
